Add resource loader (+necessary libs and helpers).
Added a new folder, asset_loaders, inside src which should contain all the code
that interacts with asset files. Also, added a new folder called assets to hold
all information related to the game that can't be in code. Incorporated
Santini's lineread library into the process of loading resources and into the
project structure with the new lib folder. Make is prepared to compile
everything. Please take care to run the game from the project's root folder, as
loaders won't be able to find assets in other cases.